Bug 12104 the test navigator shoud be able to control if feedback is visible.
There are new methods isFeedbackVisible and isReadOnly in the question base class. Questions can call these and behave appropriately.

The standard DeferredFeedback1 question should do this automatically, and only display the yellow feedback area if it is supposed to be displayed.

There is a new link in the developer servlet, to restart the question with feedback hidden, to help test questions.