Bug 12198 New API so that question authors can control scroll.
In questions you can now say scrollTo("answerbox") to scroll a
particular part of the question onto the screen, or preserveScroll(); to
keep the same scroll position when the question reloads.

Note that it only does this when the student clicks a button. If you
come back to the quetsion later, it does not automatically scroll.