Bug 11270 Fix simultaneous login to two tests.
The main symptom of this problem was having two single-question mode
questions in two iframes on the same contianing page. One would always
give an error because of the way we handled session initiaion. (Track
cookie creating by ip address, with a 1-second time-out.) Now we use a
URL paramter which is much more reliable, although the pay-off for this
is one extra redirect sometimes, when a user stars a test.