Bug 12205 support different templates for some tests.
You can now add a <templateset> tag to the deploy file. This is the name
of a subfolder of the standard template folder, where the latyout
templates should be loaded from.

Also, the location of the main template folder can be configured in
navigator.xml, although it still defaults to WEB-INF/templates.

As part of doning this, a new TemplateLoader class was created, and all
the places that need to load templates were refactored to use it.