Bug 12209 Give a clear message if the user switches between devices
We don't care if the user just leaves one device, and logs in on
another. There old session is just discarded, and we give them a new
one.

The problem comes if the student still has the test open in the first
browser window, and then tries to submit something. In that case, we
were silently creating a new session for them, and then re-displaying
the question. This certainly silently threw away their submission, and
sometimes lead to a cryptic DB error (unique key violation).

Now, they get a clear error telling them not to switch between devices.