Revise how EditAdvancedFieldComponent initialises TinyMCE #41240
The previously used solution added an extra, unnecessary, API to
QDocument / StdQuestion. Not only was that unnecessary, but also it did
not work (e.g. in Moodle with two questions on a page.)

Therefore, I re-wrote it to use a JavaScript-only solution that was
better.

This was preparatory to implementing the new SupSub editor compondent.